backend/requirements.txt
spotipy
python-dotenv
pandas
numpy
spotipy
librosa
requests
openai
matplotlib
fastapi
uvicorn
python-multipart
scikit-learn
pydantic
scipy
